ISPN-1754 IllegalStateException during cluster shutdown
Properly revert the state transfer-related part of commit d38077c.
My previous revert left the state transfer part as it was
and reverted the TransactionCoordinator.
